backend/requirements.txt
pymongo
python-dotenv
graphene
flask
flask_graphql